requirements.txt
alembic==1.5.4
aniso8601==8.1.1
appdirs==1.4.4
blurhash==1.1.4
boost==0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-pymysql==0.2.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
Mastodon.py==1.5.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
Pillow==8.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-magic==0.4.20
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.3.23
tifffile==2021.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
zipp==3.4.0
opencv-python~=4.5.1.48